+++ src/components/money-input/money-input.test.jsx
@@ -206,12 +206,6 @@ describe('MoneyInput', () => {
   describe('tooltip API', () => {
     const tooltip = 'I am an input, short and stout.';
 
-    it('applies a description to the input when the help icon is hovered', () => {
-      render(<MoneyInput {...requiredProps} tooltip={tooltip} />);
-      userEvent.hover(screen.getByRole('img', { name: 'More information' }));
-      expect(screen.getByLabelText(requiredProps.label)).toHaveAccessibleDescription(tooltip);
-    });
-
     it('removes the description to the input when the help icon is unhovered', () => {
       render(<MoneyInput {...requiredProps} tooltip={tooltip} />);
       userEvent.hover(screen.getByRole('img', { name: 'More information' }));